Move MCRAN4 to gnu
