+++ src/ivoc/ivocrand.cpp
@@ -35,6 +35,7 @@
 #include <Weibull.h>
 #include <Isaac64RNG.hpp>
 #include <NrnRandom123RNG.hpp>
+#include <MCellRan4RNG.hpp>
 
 #if HAVE_IV
 #include "ivoc.h"
@@ -63,47 +64,6 @@ static RandomPlayList* random_play_list_;
 
 #include <mcran4.h>
 
-// The decision that has to be made is whether each generator instance
-// should have its own seed or only one seed for all. We choose separate
-// seed for each but if arg not present or 0 then seed chosen by system.
-
-// the addition of ilow > 0 means that value is used for the lowindex
-// instead of the mcell_ran4_init global 32 bit lowindex.
-
-class MCellRan4: public RNG {
-  public:
-    MCellRan4(uint32_t ihigh = 0, uint32_t ilow = 0);
-    virtual ~MCellRan4();
-    virtual uint32_t asLong() {
-        return (uint32_t) (ilow_ == 0 ? mcell_iran4(&ihigh_) : nrnRan4int(&ihigh_, ilow_));
-    }
-    virtual void reset() {
-        ihigh_ = orig_;
-    }
-    virtual double asDouble() {
-        return (ilow_ == 0 ? mcell_ran4a(&ihigh_) : nrnRan4dbl(&ihigh_, ilow_));
-    }
-    uint32_t ihigh_;
-    uint32_t orig_;
-    uint32_t ilow_;
-
-  private:
-    static uint32_t cnt_;
-};
-
-MCellRan4::MCellRan4(uint32_t ihigh, uint32_t ilow) {
-    ++cnt_;
-    ilow_ = ilow;
-    ihigh_ = ihigh;
-    if (ihigh_ == 0) {
-        ihigh_ = cnt_;
-        ihigh_ = (uint32_t) asLong();
-    }
-    orig_ = ihigh_;
-}
-MCellRan4::~MCellRan4() {}
-
-uint32_t MCellRan4::cnt_ = 0;
 
 RandomPlay::RandomPlay(Rand* r, neuron::container::data_handle<double> px)
     : r_{r}
