+++ src/oc/oc_mcran4.hpp
@@ -1,0 +1,5 @@
+void set_use_mcran4(bool value);
+bool use_mcran4();
+void hoc_mcran4();
+void hoc_mcran4init();
+void hoc_usemcran4();
