+++ src/oc/scoprand.cpp
@@ -23,7 +23,8 @@ static char RCSid[] = "random.cpp,v 1.4 1999/01/04 12:46:49 hines Exp";
 #endif
 
 #include <math.h>
-#include <mcran4.h>
+#include "oc_mcran4.hpp"
+#include "mcran4.h"
 #include "scoplib.h"
 static uint32_t value = 1;
 
@@ -58,8 +59,7 @@ static uint32_t value = 1;
  *--------------------------------------------------------------------------- */
 
 double scop_random(void) {
-    extern int use_mcell_ran4_;
-    if (use_mcell_ran4_) {
+    if (use_mcran4()) {
         /*perhaps 4 times slower but much higher quality*/
         return mcell_ran4a(&value);
     } else {
